Warning検知＆修正

+++ test/test.cpp
@@ -44,7 +44,7 @@ int main(){
         );
         t_vec.push_back( th );
     }
-    for (int j=0; j<t_vec.size(); ++j){
+    for (size_t j=0; j<t_vec.size(); ++j){
         t_vec[j]->join();
     }
